% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{app_path}
\alias{app_path}
\title{Locate files in inst/app}
\usage{
app_path(...)
}
\arguments{
\item{...}{Path components relative to inst/app}
}
\value{
File path string
}
\description{
Locate files in inst/app
}
\keyword{internal}
