% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_guildparams}
\alias{default_guildparams}
\title{Default guild parameters}
\format{
A data frame with columns \code{Species}, \code{minw}, \code{maxw}, \code{Feeding.guild}.
}
\source{
Murray et al. (2022).
}
\usage{
default_guildparams
}
\description{
Table of species life stages and feeding guild classifications.
}
\keyword{datasets}
