% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_nutrition}
\alias{default_nutrition}
\title{Default nutrition parameters}
\format{
A data frame with species and nutritional content columns.
}
\source{
Scientific literature and nutritional databases.
}
\usage{
default_nutrition
}
\description{
Nutritional content values per species.
}
\keyword{datasets}
