% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{Default MizerParams object}
\format{
An object of class \code{MizerParams}.
}
\source{
Mizer project.
}
\usage{
default_params
}
\description{
A default \code{MizerParams} object for the Celtic Sea ecosystem model.
}
\keyword{datasets}
