% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_sim}
\alias{default_sim}
\title{Default MizerSim object}
\format{
An object of class \code{MizerSim}.
}
\source{
Mizer project.
}
\usage{
default_sim
}
\description{
A default \code{MizerSim} object representing a 10-year baseline projection.
}
\keyword{datasets}
