% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrition.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize nutrition data on package load}
\usage{
.onLoad(libname, pkgname)
}
\description{
Populates internal datasets \code{nut} and \code{all_matches} used by nutrition plots.
Not part of the public API.
}
\keyword{internal}
