% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_guilds.R
\name{guildplot}
\alias{guildplot}
\title{Plot guild biomass change}
\usage{
guildplot(
  harvestedprojection,
  unharvestedprojection,
  chosenyear,
  guildparams,
  celticsim,
  mode = c("chosen", "triple")
)
}
\description{
Plot guild biomass change
}
\keyword{internal}
