% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_noninteracting.R
\name{make_noninteracting}
\alias{make_noninteracting}
\title{Make model non-interacting}
\usage{
make_noninteracting(params)
}
\arguments{
\item{params}{A MizerParams object}
}
\value{
The modified MizerParams object
}
\description{
Adds the predation mortality to the external mortality and the predation
encounter to the external encounter and then sets the interaction matrix to
zero and switches off the interaction with the resource, so that the model
becomes non-interacting.
}
