% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerShiny-package.R
\docType{package}
\name{mizerShiny-package}
\alias{mizerShiny-package}
\title{mizerShiny: mizerShiny}
\description{
Frontend to running mizer models.
}
\author{
\strong{Maintainer}: Luca Broadbent \email{luke.broadbent@cefas.gov.uk}

Authors:
\itemize{
  \item Luca Broadbent \email{luke.broadbent@cefas.gov.uk}
}

}
\keyword{internal}
