% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{mizerShiny}
\alias{mizerShiny}
\title{Launch mizerShiny}
\usage{
mizerShiny(
  params = default_params,
  guildparams = default_guildparams,
  nutrition = default_nutrition,
  fishery_strategy_tabs = c("Biomass", "Biomass \% Change", "Yield", "Yield \% Change",
    "Nutrition", "Length", "Guild"),
  species_role_tabs = c("Biomass", "Size", "Guilds", "Diet"),
  ...
)
}
\arguments{
\item{params}{A \link[mizer:MizerParams]{mizer::MizerParams} object. Defaults to \code{default_params}.}

\item{guildparams}{An optional data frame defining guilds for the model.
See the 'Guilds' section of the package vignette for details.}

\item{nutrition}{An optional data frame defining nutritional values
for the model. See the 'Nutrition' section of the package vignette for details.}

\item{fishery_strategy_tabs}{A character vector specifying which tabs to include
in the Fishery Strategy module. Tab names that don't correspond to existing
tabs will be ignored.}

\item{species_role_tabs}{A character vector specifying which tabs to include
in the Species Role module. Tab names that don't correspond to existing
tabs will be ignored.}

\item{...}{further arguments passed on to \code{shiny::runApp()}
(e.g. \code{host}, \code{port}, \code{launch.browser}).}
}
\description{
Launch mizerShiny
}
\examples{
\dontrun{
# Run with default model
mizerShiny()

# Run with custom model
mizerShiny(NS_params)

# Run with only specific tabs
mizerShiny(fishery_strategy_tabs = c("Biomass", "Yield"),
           species_role_tabs = c("Biomass", "Diet"))
}
}
