% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrition.R
\name{normalise_nutrition_cols}
\alias{normalise_nutrition_cols}
\title{Normalise nutrition column names}
\usage{
normalise_nutrition_cols(nutrition)
}
\arguments{
\item{nutrition}{Data frame of per-species nutrient contents}
}
\value{
Data frame with normalised column names
}
\description{
Alter nutrient column names to be more readable and drop unwanted nutrients.
}
\keyword{internal}
