% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrition.R
\name{plotNutrition}
\alias{plotNutrition}
\title{Plot relative nutritional output compared to a reference simulation}
\usage{
plotNutrition(sims, ref, step, nutrition = default_nutrition)
}
\arguments{
\item{sims}{List of mizer projection objects to compare}

\item{ref}{Reference mizer projection}

\item{step}{Integer time index to evaluate}

\item{nutrition}{Data frame of per-species nutrient contents}
}
\value{
A ggplot object
}
\description{
Aggregates nutrient yields from simulations at a given time \code{step} and
compares each to a reference, returning a grouped bar plot of percentage
differences per nutrient.
}
\keyword{internal}
