% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plotSpeciesActualBiomass}
\alias{plotSpeciesActualBiomass}
\title{Plot species actual biomass at selected times}
\usage{
plotSpeciesActualBiomass(
  harvestedprojection,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\arguments{
\item{harvestedprojection}{Harvested mizer projection}

\item{chosenyear}{Integer defining the full period; quarter/half derived}

\item{mode}{Either "triple" or "chosen"}
}
\value{
A ggplot object
}
\description{
Renders a bar chart of species actual biomass values at selected times for
a harvested projection.
}
\keyword{internal}
