% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plotSpeciesActualYield}
\alias{plotSpeciesActualYield}
\title{Plot species actual yield at selected times}
\usage{
plotSpeciesActualYield(
  harvestedprojection,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\arguments{
\item{harvestedprojection}{Harvested mizer projection}

\item{chosenyear}{Integer defining the full period; quarter/half derived}

\item{mode}{Either "triple" or "chosen"}
}
\value{
A ggplot object
}
\description{
Renders a stacked bar chart of species actual yield values by gear at selected times for
a harvested projection.
}
\keyword{internal}
