% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plotSpeciesWithTimeRange}
\alias{plotSpeciesWithTimeRange}
\title{Plot species biomass change at selected times}
\usage{
plotSpeciesWithTimeRange(
  harvestedprojection,
  unharvestedprojection,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\description{
Plot species biomass change at selected times
}
\keyword{internal}
