% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_2sims.R
\name{plotSpeciesWithTimeRange2}
\alias{plotSpeciesWithTimeRange2}
\title{Plot species biomass change for two simulations}
\usage{
plotSpeciesWithTimeRange2(
  harvestedprojection1,
  harvestedprojection2,
  unharvestedprojection,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\description{
Plot species biomass change for two simulations
}
\keyword{internal}
