% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{plotSpeciesYieldChange}
\alias{plotSpeciesYieldChange}
\title{Plot species yield change at selected times}
\usage{
plotSpeciesYieldChange(
  harvestedprojection,
  sim_0,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\arguments{
\item{harvestedprojection}{Harvested mizer projection}

\item{sim_0}{Unharvested mizer projection}

\item{chosenyear}{Integer defining the full period; quarter/half derived}

\item{mode}{Either "triple" or "chosen"}
}
\value{
A ggplot object
}
\description{
Renders a bar chart of species yield percentage change at selected times for
a single harvested vs unharvested comparison.
}
\keyword{internal}
