% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{plotSpectraRelative}
\alias{plotSpectraRelative}
\title{Relative community size spectrum (single vs baseline)}
\usage{
plotSpectraRelative(object1, object2, time1, time2)
}
\description{
Relative community size spectrum (single vs baseline)
}
\keyword{internal}
