% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species.R
\name{process_sim_shared}
\alias{process_sim_shared}
\title{Shared processing for species biomass plots}
\usage{
process_sim_shared(
  harvestedprojection,
  unharvestedprojection,
  chosenyear,
  mode = c("triple", "chosen")
)
}
\description{
Shared processing for species biomass plots
}
\keyword{internal}
