% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{setupYearControls}
\alias{setupYearControls}
\title{Setup +/- year controls associated with a slider}
\usage{
setupYearControls(input, session, sliderId, minusId, plusId)
}
\arguments{
\item{input}{Shiny input}

\item{session}{Shiny session}

\item{sliderId}{Id of the slider input}

\item{minusId}{Id of the "-" button}

\item{plusId}{Id of the "+" button}
}
\value{
Invisibly, NULL
}
\description{
Attaches observers to increment/decrement a slider input and dynamically cap
its maximum to twice the selected value within the range from 12 to 100.
}
\keyword{internal}
