% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrition.R
\name{calc_nutrition_totals}
\alias{calc_nutrition_totals}
\title{Calculate total nutrients from a simulation at given time steps}
\usage{
calc_nutrition_totals(sim, steps)
}
\arguments{
\item{sim}{Mizer projection object}

\item{steps}{Integer time indices to evaluate (can be a range)}
}
\value{
Named vector of nutrient totals
}
\description{
Helper function to compute nutrient totals from yield data.
}
\keyword{internal}
