% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_2sims.R
\name{guildplot_both}
\alias{guildplot_both}
\title{Plot guild biomass change for two simulations}
\usage{
guildplot_both(
  harvestedprojection1,
  harvestedprojection2,
  unharvestedprojection,
  chosenyear,
  guildparams,
  celticsim,
  mode = c("chosen", "triple")
)
}
\description{
Plot guild biomass change for two simulations
}
\keyword{internal}
