% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{mizerShiny}
\alias{mizerShiny}
\title{Launch mizerShiny}
\usage{
mizerShiny(params = default_params, ...)
}
\arguments{
\item{params}{A MizerParams object. Defaults to NS_params.}

\item{...}{further arguments passed on to \code{shiny::runApp()}
(e.g. \code{host}, \code{port}, \code{launch.browser}).}
}
\description{
Launch mizerShiny
}
\examples{
\dontrun{
# Run with default model
mizerShiny()

# Run with custom model
mizerShiny(NS_params)
}
}
