% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nutrition.R
\name{normalise_nutrition_cols}
\alias{normalise_nutrition_cols}
\title{Calculate total nutrients from a simulation at given time steps}
\usage{
normalise_nutrition_cols(nutrition)
}
\arguments{
\item{nutrition}{Data frame of per-species nutrient contents}

\item{sim}{Mizer projection object}

\item{steps}{Integer time indices to evaluate (can be a range)}
}
\value{
Named vector of nutrient totals
}
\description{
Helper function to compute nutrient totals from yield data.
}
\keyword{internal}
