% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentdiff.R
\name{percentdiff}
\alias{percentdiff}
\title{Percentage difference between two values by species}
\usage{
percentdiff(df, df_0)
}
\arguments{
\item{df}{Data frame with columns \code{Species} and \code{value}}

\item{df_0}{Data frame with columns \code{Species} and \code{value} for baseline}
}
\value{
A data frame with columns \code{Species} and \code{percentage_diff}
}
\description{
Joins two data frames by \code{Species} and returns percentage change of
values in first data frame relative to the second, excluding the \code{Resource} row.
}
\examples{
percentdiff(
  df = data.frame(Species = c("A","B"), value = c(120, 80)),
  df_0 = data.frame(Species = c("A","B"), value = c(100, 100))
)
}
