% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectra.R
\name{plotSpectraRelative2}
\alias{plotSpectraRelative2}
\title{Relative spectra for two sims vs baseline}
\usage{
plotSpectraRelative2(object1, object2, object3, time1, time2)
}
\description{
Relative spectra for two sims vs baseline
}
\keyword{internal}
