% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch.R
\name{plotYieldVsSize}
\alias{plotYieldVsSize}
\alias{plotlyYieldVsSize}
\title{Plot the size distribution of the catch}
\usage{
plotYieldVsSize(
  sim,
  species = NULL,
  gear = NULL,
  x_var = c("Weight", "Length"),
  return_data = FALSE
)

plotlyYieldVsSize(
  sim,
  species = NULL,
  gear = NULL,
  x_var = c("Weight", "Length"),
  ...
)
}
\arguments{
\item{sim}{An object of class \link[mizer:MizerSim-class]{MizerSim} or list of objects}

\item{species}{The name of the predator species for which to plot the
mortality.}

\item{gear}{Optional. The name of a gear. If supplied, only the yield from
this gear will be displayed.}

\item{x_var}{Determines whether to show the size distribution of the catch as
a function of weight ("Weight") or as a function of length ("Length").
Default is "Weight".}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is
FALSE}

\item{...}{Other arguments passed to \code{plotYieldVsSize}}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a
data frame with the four
variables 'w' or 'l' (depending on \code{x_var}), 'Catch density', 'Type', 'Species
and the second slot is a data frame with the four variables 'w_mat',
'Species', 'y_coord', 'Type' (to plot vertical lines).
}
\description{
Plots the normalised number density of the catch for a species as a function
of either length or weight.
Superimposes a plot of the number density of all individuals of the
species.
}
\seealso{
\link[mizer:plotting_functions]{mizer::plotting_functions}
}
\concept{plotting functions}
