% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{setupYearControls}
\alias{setupYearControls}
\title{Setup +/- year controls associated with a slider}
\usage{
setupYearControls(input, session, sliderId, minusId, plusId)
}
\description{
Setup +/- year controls associated with a slider
}
\keyword{internal}
